Widen MetadataRecord.interval from int to float
DB column is `numeric NOT NULL`, but the Pydantic model rejected anything
non-integer. Streaming sensors (e.g. RFObserver) want sub-second values
like 0.25 / 0.5 to describe continuous capture cadence; the schema
already accepts them.

Backward compatible: existing iq-feeder traffic sends integer seconds
which still validate as float.

Bumps version to 0.2.3.

+++ src/rf_shared/models.py
@@ -64,7 +64,7 @@ class MetadataRecord(BaseModel):
 
     # Radio settings
     frequency: int
-    interval: int
+    interval: float
     length: float
     gain: int
     sampling_rate: int
